backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai>=0.3.0
Pillow>=10.0.0
PyPDF2>=3.0.0
resend>=0.7.0
slowapi>=0.1.9
lima>=0.5.0
orjson>=3.9.0
psutil>=5.9.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
